+++ app/assets/stylesheets/deliveries.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the deliverys controller here.
+// Place all the styles related to the deliveries controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
