+++ app/helpers/deliveries_helper.rb
@@ -1,0 +1,2 @@
+module DeliveriesHelper
+end
