--- app/helpers/deliverys_helper.rb
@@ -1,2 +1,0 @@
-module DeliverysHelper
-end
