+++ app/models/prefecture.rb
@@ -52,4 +52,5 @@ class Prefecture < ActiveHash::Base
 
   include ActiveHash::Associations
   belongs_to :items
+  belongs_to :deliverys
 end
