+++ app/views/items/show.html.erb
@@ -28,7 +28,7 @@
         <p class="or-text">or</p>
         <%= link_to "削除", item_path(@item), method: :delete, class:"item-destroy" %>
       <% elsif user_signed_in? %>
-        <%= link_to "購入画面に進む", deliverys_path,class:"item-red-btn"%>
+        <%= link_to "購入画面に進む", item_deliveries_path(@item),class:"item-red-btn"%>
       <% end %>
 
     <div class="item-explain-box">
