+++ spec/helpers/deliveries_helper_spec.rb
@@ -1,15 +1,15 @@
 require 'rails_helper'
 
 # Specs in this file have access to a helper object that includes
-# the DeliverysHelper. For example:
+# the DeliveriesHelper. For example:
 #
-# describe DeliverysHelper do
+# describe DeliveriesHelper do
 #   describe "string concat" do
 #     it "concats two strings with spaces" do
 #       expect(helper.concat_strings("this","that")).to eq("this that")
 #     end
 #   end
 # end
-RSpec.describe DeliverysHelper, type: :helper do
+RSpec.describe DeliveriesHelper, type: :helper do
   pending "add some examples to (or delete) #{__FILE__}"
 end
