+++ spec/requests/deliveries_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Deliveries", type: :request do
+
+end
