--- spec/requests/deliverys_request_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Deliverys", type: :request do
-
-end
